+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/mapping/impl/AnnotationHandlerNotFoundException.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.servlet.mvc.mapping.impl;
 
 public class AnnotationHandlerNotFoundException extends RuntimeException {
