+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/mapping/impl/ControllerScanner.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.servlet.mvc.mapping.impl;
 
 import com.interface21.context.stereotype.Controller;
 import java.lang.reflect.Constructor;
